pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "gamma"
version = "0.0.1"
authors = [
  { name="Shuheng Liao" },
  { name="Ashkan Golgoon" },
]
description = "A GPU-accelerated FEM solver for residual stress simulation in additive manufacturing using CuPy"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "botorch>=0.16.1",
    "cupy-cuda12x>=14.0.0",
    "ipykernel>=7.2.0",
    "numba>=0.63.1",
    "pandas>=2.3.3",
    "pytest>=9.0.2",
    "pyvista>=0.47.0",
    "zarr>=2.18.3",
]

[project.urls]
"Homepage" = "https://github.com/ShuhengLiao/AM_Thermomechanical_Solver"
"Bug Tracker" = "https://github.com/ShuhengLiao/AM_Thermomechanical_Solver/issues"
